Adicionando mais validações

+++ src/main/java/com/example/pdv/service/VendaService.java
@@ -7,6 +7,7 @@ import com.example.pdv.domain.Venda;
 import com.example.pdv.dto.ItemVendaDTO;
 import com.example.pdv.dto.VendaDTO;
 import com.example.pdv.repository.VendaRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -28,6 +29,7 @@ public class VendaService {
     public Venda insert(VendaDTO dto){
         Venda venda = new Venda(dto);
 
+        //Todo: talvez devessemos deixar a parte de gravar os itens venda em um Service diferente e tirar a cascata.
         List<ItemVenda> listaItens = new ArrayList<>();
         for(ItemVendaDTO itemDto : dto.getItensVenda()){
             Produto produto = produtoService.findById(itemDto.getProdutoId());
@@ -49,18 +51,33 @@ public class VendaService {
     }
 
     public Venda update(Venda venda){
+        if(!vendaRepository.existsById(venda.getId()))
+            throw new EntityNotFoundException("Venda id " + venda.getId() + " não encontrada.");
         return vendaRepository.save(venda);
     }
 
     public List<Venda> findAll(){
-        return vendaRepository.findAll();
+        List<Venda> vendas = vendaRepository.findAll();
+
+        if(vendas.isEmpty())
+            throw new EntityNotFoundException("Não foram encontradas vendas.");
+
+        return vendas;
     }
 
-    public Optional<Venda> findById(Integer id){
-        return vendaRepository.findById(id);
+    public Venda findById(Integer id){
+        Optional<Venda>  venda =  vendaRepository.findById(id);
+
+        if(venda.isEmpty())
+            throw new EntityNotFoundException("Venda id " + id + " não encontrada.");
+
+        return venda.get();
     }
 
     public void delete(Integer id){
+        if(!vendaRepository.existsById(id))
+            throw new EntityNotFoundException("Venda id " + id + " não encontrada.");
+
         vendaRepository.deleteById(id);
     }
 }
